requirements.txt
aiofiles==23.2.1
asgiref==3.8.1
bcrypt==4.2.1
cffi==1.17.1
channels==4.2.0
channels-redis==4.0.0
chardet==5.2.0
daphne==4.1.2
Django
djangorestframework==3.15.2
future==1.0.0
Jinja2==3.1.5
netmiko==4.5.0
ntc_templates==7.7.0
paramiko>=2.8.1
rich==13.9.4
setuptools==75.8.0
textfsm==1.1.3
django-redis>=5.2.0
redis>=4.0.0
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
mysqlclient==2.2.4